Storing names with using "\n" and "\t"
So basically this is my work of 6th assignment and  I was given this, 
Stripping Names: Store a person’s name, and include some whitespace characters at the beginning and end of the name. Make sure you use each character combination, "\t" and "\n", at least once. Print the name once, so the whitespace around the name is displayed. Then print the name after striping the white spaces.
YOU CAN ALSO GIVE THIS A TRY.